refactor: create one service for each entity
